# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RescuEat API - Food rescue platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List all active stores with available bags
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          description: Radius in km
          schema:
            type: number
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a new store (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{storeId}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get store details
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateStore
      tags: [stores]
      summary: Update store info
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreRequest"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Store:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        category:
          type: string
          enum: [restaurant, bakery, cafe, supermarket, convenience, other]
        lat:
          type: number
        lng:
          type: number
        imageUrl:
          type: string
        phone:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        rating:
          type: number
        isActive:
          type: boolean
        ownerId:
          type: string
        totalBagsAvailable:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - address
        - city
        - category
        - lat
        - lng
        - isActive
        - totalBagsAvailable
        - createdAt
    CreateStoreRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        category:
          type: string
          enum: [restaurant, bakery, cafe, supermarket, convenience, other]
        lat:
          type: number
        lng:
          type: number
        imageUrl:
          type: string
        phone:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
      required:
        - name
        - address
        - city
        - category
        - lat
        - lng
    UpdateStoreRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        category:
          type: string
          enum: [restaurant, bakery, cafe, supermarket, convenience, other]
        imageUrl:
          type: string
        phone:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        isActive:
          type: boolean
...