# lib/api-zod/src/generated/types/updateStoreRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RescuEat API - Food rescue platform
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateStoreRequestCategory } from "./updateStoreRequestCategory";

export interface UpdateStoreRequest {
  name?: string;
  description?: string;
  address?: string;
  city?: string;
  category?: UpdateStoreRequestCategory;
  imageUrl?: string;
  phone?: string;
  openTime?: string;
  closeTime?: string;
  isActive?: boolean;
}